Refactor createPaintStyle with React hooks
