+++ src/index.ts
@@ -22,4 +22,4 @@ export { Slice } from './components/slice/Slice';
 export { StyleSheet } from './helpers/StyleSheet';
 export { Platform } from './helpers/Platform';
 
-export { createPaintStyle } from './localStyles/createPaintStyle';
+export { usePaintStyle } from './localStyles/usePaintStyle';
